+++ VideoPlayer/VideoPlayer.js
@@ -35,6 +35,7 @@ import {isValidElement, cloneElement, Component} from 'react';
 import ReactDOM from 'react-dom';
 
 import $L from '../internal/$L';
+import Button from '../Button';
 import Skinnable from '../Skinnable';
 import Spinner from '../Spinner';
 import {
@@ -84,7 +85,7 @@ const RootContainer = SpotlightContainerDecorator(
 
 const ControlsContainer = SpotlightContainerDecorator(
 	{
-		enterTo: '',
+		enterTo: 'default-element',
 		straightOnly: true
 	},
 	'div'
@@ -399,6 +400,14 @@ const VideoPlayerBase = class extends Component {
 		 * @public
 		 */
 		noSpinner: PropTypes.bool,
+
+		/**
+		 * Called when the back button is clicked.
+		 *
+		 * @type {Function}
+		 * @public
+		 */
+		onBack: PropTypes.func,
 
 		/**
 		 * Called when the player's controls change availability, whether they are shown
@@ -1744,11 +1753,6 @@ const VideoPlayerBase = class extends Component {
 		};
 	};
 
-	disablePointerMode = () => {
-		Spotlight.setPointerMode(false);
-		return true;
-	};
-
 	//
 	// Player Interaction events
 	//
@@ -1836,8 +1840,6 @@ const VideoPlayerBase = class extends Component {
 			}
 		} else if (is('up', keyCode)) {
 			Spotlight.setPointerMode(false);
-			preventDefault(ev);
-			stopImmediate(ev);
 		}
 	};
 
@@ -1931,6 +1933,7 @@ const VideoPlayerBase = class extends Component {
 			noMiniFeedback,
 			noSlider,
 			noSpinner,
+			onBack,
 			selection,
 			spotlightDisabled,
 			spotlightId,
@@ -2031,6 +2034,16 @@ const VideoPlayerBase = class extends Component {
 						>
 							{secondsToTime(this.state.sliderTooltipTime, durFmt)}
 						</FeedbackContent>
+						{
+							this.state.mediaControlsVisible ?
+								<Button
+									className={css.back}
+									icon="arrowhookleft"
+									onClick={onBack}
+									size="small"
+								/> :
+								null
+						}
 						<ControlsContainer
 							className={css.bottom + (this.state.mediaControlsVisible ? '' : ' ' + css.hidden) + (this.state.infoVisible ? ' ' + css.lift : '')}
 							spotlightDisabled={spotlightDisabled || !this.state.mediaControlsVisible}
@@ -2074,7 +2087,6 @@ const VideoPlayerBase = class extends Component {
 										onFocus={this.handleSliderFocus}
 										onKeyDown={this.handleSliderKeyDown}
 										onKnobMove={this.handleKnobMove}
-										onSpotlightUp={this.handleSpotlightUpFromSlider}
 										selection={proportionSelection}
 										spotlightDisabled={spotlightDisabled || !this.state.mediaControlsVisible}
 										value={this.state.proportionPlayed}