+++ samples/sampler/stories/default/VideoPlayer.js
@@ -53,6 +53,7 @@ const prop = {
 	},
 	events: [
 		'onAbort',
+		'onBack',
 		'onCanPlay',
 		'onCanPlayThrough',
 		'onControlsAvailable',